fix: #113 Create Notification Manager
