+++ TMT/TMT/BusJourneyScanned/Views/ScannedJourneyInfoView.swift
@@ -137,7 +137,7 @@ struct ScannedJourneyInfoView: View {
                             
                             guard let startStop = journeyModel.journeyStops.first else { return }
                             guard let endStop = journeyModel.journeyStops.last else { return }
-
+                            
                             activityManager.startLiveActivity(startBusStop: startStop, endBusStop: endStop, remainingStops: locationManager.remainingStops)
                             tag = 1
                             path.append("BusStop")
